refactor: decompose progress module for SoC (fixes #148)
Split reporter.rs (446 LOC) into focused modules:
- mod.rs: ProgressEvent + utils + re-exports (80 LOC)
- emitter.rs: ProgressEmitter Tauri transport (204 LOC)
- state.rs: ProgressReporter state machine (107 LOC)

Changes:
- Promote format_eta/calculate_stage_progress to free functions
- Consolidate event.rs + utils.rs into mod.rs (KISS)
- Remove _stage_weight dead code from calculate_progress()
- Rename reporter.rs → state.rs for clarity
- Migrate inline tests to external test file using public APIs
- Update mod.rs re-exports to preserve public API stability
- Update audio/mod.rs to export util functions
- Update doc references (events.ts, architecture-audit)

All modules now under 400 LOC limit with high cohesion.

Verified:
- cargo fmt/clippy: pass
- cargo test: 106 tests pass
- ensure-contract.sh: pass
- bun run build: pass

/gemini review
